[SQL] Make 'emitHandles' an internal compiler options
