+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/BaseSQLTests.java
@@ -157,6 +157,7 @@ public class BaseSQLTests {
         options.languageOptions.throwOnError = true;
         options.languageOptions.generateInputForEveryTable = true;
         options.ioOptions.quiet = true;
+        options.ioOptions.emitHandles = false;
         options.languageOptions.incrementalize = incremental;
         options.languageOptions.optimizationLevel = optimize ? 2 : 1;
         return options;
