+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/suites/TpchTest.java
@@ -1,19 +1,16 @@
 package org.dbsp.sqlCompiler.compiler.sql.suites;
 
-import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
 import org.dbsp.sqlCompiler.compiler.CompilerOptions;
 import org.dbsp.sqlCompiler.compiler.sql.BaseSQLTests;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.TestUtil;
-import org.dbsp.sqlCompiler.compiler.backend.rust.RustFileWriter;
-import org.dbsp.util.Utilities;
 import org.junit.Test;
 
 import java.io.IOException;
 
 public class TpchTest extends BaseSQLTests {
     @Test
-    public void compileTpch() throws IOException, InterruptedException {
+    public void compileTpch() throws IOException {
         String tpch = TestUtil.readStringFromResourceFile("tpch.sql");
         CompilerOptions options = this.testOptions(true, true);
         DBSPCompiler compiler = new DBSPCompiler(options);
@@ -22,10 +19,6 @@ public class TpchTest extends BaseSQLTests {
         compiler.compileStatements(tpch);
         System.err.println(compiler.messages);
         compiler.throwIfErrorsOccurred();
-        DBSPCircuit circuit = getCircuit(compiler);
-        RustFileWriter writer = new RustFileWriter(compiler, testFilePath);
-        writer.add(circuit);
-        writer.writeAndClose();
-        Utilities.compileAndTestRust(rustDirectory, true);
+        this.addRustTestCase("tpch", compiler, getCircuit(compiler));
     }
 }
